#14 : Si(ng) renamed to Si(gn)
