+++ src/App.jsx
@@ -4,8 +4,8 @@ import './env.css';
 import './App.css';
 
 import Home from './pages/Home/Home';
-import SingIn from './pages/SingIn/SingIn';
-import SingUp from './pages/SingUp/SingUp';
+import SignIn from './pages/SignIn/SignIn';
+import SignUp from './pages/SignUp/SignUp';
 import Task from './pages/Task/Task';
 import Error404 from './pages/Error404/Error404';
 
@@ -13,8 +13,8 @@ function App() {
   return (
     <Routes>
       <Route path="/" element={<Home />} />
-      <Route path="/sing-in" element={<SingIn />} />
-      <Route path="/sing-up" element={<SingUp />} />
+      <Route path="/sign-in" element={<SignIn />} />
+      <Route path="/sign-up" element={<SignUp />} />
       <Route path="/task" element={<Task />} />
       <Route path="*" element={<Error404 />} />
     </Routes>
